Add option flags to install script

+++ install.sh
@@ -1,7 +1,60 @@
 #/bin/sh
 
-NAMESPACE=$(helm inspect kiam-install/|grep namespace|cut -d: -f2|xargs)
-SERVICENAME=$(helm inspect kiam-install/|grep serviceName|cut -d: -f2|xargs)
+function usage(){
+  cat<<EOF
+Install Kiam, generating certs if requested.
+  Usage:  $0 [OPTIONS]
+
+  Options:
+    -d   Directory to install chart from
+    -c   Generate certs and save them as secrets
+    -r   Release name for helm to use
+EOF
+}
+
+while [[ ! -z $1 ]]; do
+  case $1 in
+    -d|--tag)
+      shift
+      DIR="$1"
+      ;;
+    -c|--generate-certs)
+      CERTS="true"
+      ;;
+    -r|--release)
+      shift
+      RELEASE="$1"
+      ;;
+    *)
+      usage
+      exit 1
+      ;;
+  esac
+  shift
+done
+
+if [ -z $DIR ] || [ -z $RELEASE ]; then
+  echo "-d and -r options are required"
+  usage
+  exit 2
+fi
+
+if ! helm lint $DIR >/dev/null 2>&1; then
+  echo "$DIR is not a path to a valid helm chart"
+  usage
+  exit 2
+fi
+
+if [ ! -f ${DIR}/values.yaml ]; then
+  echo "No values.yaml file included in helm chart"
+  usage
+  exit 2
+fi
+
+NAMESPACE="-n $(helm inspect $DIR |grep namespace|cut -d: -f2|xargs)"
+SERVICENAME=$(helm inspect $DIR |grep serviceName|cut -d: -f2|xargs)
+AGENT_TLS=$(helm inspect $DIR |grep agentTls|cut -d: -f2|xargs)
+SERVER_TLS=$(helm inspect $DIR |grep serverTls|cut -d: -f2|xargs)
 FQDN=${SERVICENAME}.${NAMESPACE}.svc.cluster.local
 
 JSON=$(cat<<EOF
@@ -34,26 +87,28 @@ JSON=$(cat<<EOF
 EOF
 )
 
-echo "$JSON" >server.json
+if [ ! -z $CERTS ]; then
+  echo "$JSON" >server.json
 
-cfssl gencert -initca ca.json | cfssljson -bare ca
-cfssl gencert -ca=ca.pem -ca-key=ca-key.pem -config=ca-config.json -profile=default server.json | cfssljson -bare server
-cfssl gencert -ca=ca.pem -ca-key=ca-key.pem -config=ca-config.json -profile=default agent.json | cfssljson -bare agent
+  cfssl gencert -initca ca.json | cfssljson -bare ca
+  cfssl gencert -ca=ca.pem -ca-key=ca-key.pem -config=ca-config.json -profile=default server.json | cfssljson -bare server
+  cfssl gencert -ca=ca.pem -ca-key=ca-key.pem -config=ca-config.json -profile=default agent.json | cfssljson -bare agent
 
-kubectl --kubeconfig ~/.kube/config delete secret kiam-server-tls $NAMESPACE >/dev/null 2>&1
-kubectl --kubeconfig ~/.kube/config delete secret kiam-agent-tls $NAMESPACE >/dev/null 2>&1
+  kubectl --kubeconfig ~/.kube/config delete secret $AGENT_TLS $NAMESPACE >/dev/null 2>&1
+  kubectl --kubeconfig ~/.kube/config delete secret $SERVER_TLS $NAMESPACE >/dev/null 2>&1
 
-kubectl --kubeconfig ~/.kube/config create secret generic $NAMESPACE kiam-server-tls  \
-  --from-file=ca.pem \
-  --from-file=server.pem \
-  --from-file=server-key.pem
+  kubectl --kubeconfig ~/.kube/config create secret generic $NAMESPACE $SERVER_TLS  \
+    --from-file=ca.pem \
+    --from-file=server.pem \
+    --from-file=server-key.pem
 
-kubectl --kubeconfig ~/.kube/config create secret generic $NAMESPACE kiam-agent-tls  \
-  --from-file=ca.pem \
-  --from-file=agent.pem \
-  --from-file=agent-key.pem
+  kubectl --kubeconfig ~/.kube/config create secret generic $NAMESPACE $AGENT_TLS  \
+    --from-file=ca.pem \
+    --from-file=agent.pem \
+    --from-file=agent-key.pem
 
-helm del --purge ${1}
-helm install -f ${1}/values.yaml ./${1}/ --name ${2}
+    rm -f *.pem
+fi
 
-rm -f *.pem
+helm del --purge ${RELEASE}
+helm install -f ${DIR}/values.yaml ./${DIR}/ --name ${RELEASE}
